min up/down time with variable step length Symbolic controller
Now it´s possible to dicrectly define a minimum up or downtime in minutes and use different step length. if the minimum time is not integer divisible by the step time, the minumum steps will be round up. For example if the step ist 30 min and the minimum uptime for a unit is 61 min, the unit will be held up for 90 minutes which equals 3 steps